+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _Unreleased_
 
 - Fixed "unauthorized" error which occurred while updating email and password at
   same time.
+- Improve message for `machines list` when no machines are found.
 
 ## v0.17.0